Fix compilation errors and warnings
- Fixed missing constructor parameters in test mocks (IMetricsRecorder, IHttpContextAccessor, IFeatureManager, ILogger, IRegulatoryRulesEngine, IPaymentStateService, IIncidentReportGenerator)
- Fixed DTO/record constructor calls (ThreeDSecureResultDto, SplitRuleDto, PaymentProviderInfoDto, CreatePaymentDto)
- Fixed method signature mismatches (UpdateStatusAsync -> UpdatePaymentStatus, AddAsync return types)
- Resolved namespace ambiguities (Payment.Domain.Common.Result vs GreenDonut.Result)
- Fixed ActionResult<T> test assertions to use .Result property
- Fixed nullable reference type warnings (CS8625) using null! for intentional nulls
- Fixed async method warnings (CS1998) by removing async/using Task.FromResult
- Fixed xUnit2002 warnings by replacing Assert.NotNull on value types
- Fixed ASP0000 warning with pragma for BuildServiceProvider in Swagger config
- Added missing using statements and package references
- Commented out AddProcessInstrumentation (not available in current OpenTelemetry version)
- Commented out UIResponseWriter usage due to namespace resolution issues

Build now succeeds with 0 warnings and 0 errors.

+++ src/Payment.Infrastructure/Services/AnomalyDetectionService.cs
@@ -49,7 +49,7 @@ public class AnomalyDetectionService : IAnomalyDetectionService
                 anomalies.Add(new
                 {
                     DetectedAt = point.Date,
-                    Type = point.TotalProcessed > threshold ? "Volume Spike" : "Volume Drop",
+                    Type = point.TotalProcessed > (decimal)threshold ? "Volume Spike" : "Volume Drop",
                     Severity = severity,
                     Description = $"Unusual transaction volume detected: {point.TotalProcessed:F2} (expected range: {lowerThreshold:F2} - {threshold:F2})",
                     ExpectedValue = (decimal)mean,
